+++ src/main/java/com/doddi/vulnerability_scanner_idea_plugin/config/ApplicationSettings.java
@@ -4,9 +4,9 @@ import java.util.HashMap;
 import java.util.Map;
 
 import com.doddi.vulnerability_scanner_idea_plugin.service.ScannerService;
-import com.doddi.vulnerability_scanner_idea_plugin.service.scanners.DummyUIConfiguration;
-import com.doddi.vulnerability_scanner_idea_plugin.service.scanners.NexusIQUIConfiguration;
-import com.doddi.vulnerability_scanner_idea_plugin.service.scanners.OssIndexUIConfiguration;
+import com.doddi.vulnerability_scanner_idea_plugin.service.scanners.dummy.DummyUIConfiguration;
+import com.doddi.vulnerability_scanner_idea_plugin.service.scanners.nexusiq.NexusIQUIConfiguration;
+import com.doddi.vulnerability_scanner_idea_plugin.service.scanners.ossindex.OssIndexUIConfiguration;
 import com.google.common.base.Strings;
 import com.intellij.credentialStore.CredentialAttributes;
 import com.intellij.credentialStore.CredentialAttributesKt;
