+++ src/main/java/com/doddi/vulnerability_scanner_idea_plugin/config/ScannerTypeComboBoxModel.java
@@ -2,7 +2,7 @@ package com.doddi.vulnerability_scanner_idea_plugin.config;
 
 import javax.swing.*;
 
-import com.doddi.vulnerability_scanner_idea_plugin.service.scanners.OssIndexUIConfiguration;
+import com.doddi.vulnerability_scanner_idea_plugin.service.scanners.ossindex.OssIndexUIConfiguration;
 
 public class ScannerTypeComboBoxModel
     extends DefaultComboBoxModel<UIConfiguration>
