+++ src/main/java/com/doddi/vulnerability_scanner_idea_plugin/service/scanners/dummy/DummyUIConfiguration.java
@@ -1,4 +1,4 @@
-package com.doddi.vulnerability_scanner_idea_plugin.service.scanners;
+package com.doddi.vulnerability_scanner_idea_plugin.service.scanners.dummy;
 
 import com.doddi.vulnerability_scanner_idea_plugin.config.UIConfiguration;
 import com.doddi.vulnerability_scanner_idea_plugin.service.ScannerService;
