+++ src/main/java/com/doddi/vulnerability_scanner_idea_plugin/service/scanners/nexusiq/NexusIQScannerService.java
@@ -1,4 +1,4 @@
-package com.doddi.vulnerability_scanner_idea_plugin.service.scanners;
+package com.doddi.vulnerability_scanner_idea_plugin.service.scanners.nexusiq;
 
 import java.net.URI;
 import java.util.Arrays;
