+++ src/main/java/com/doddi/vulnerability_scanner_idea_plugin/service/scanners/nexusiq/NexusIQUIConfiguration.java
@@ -1,26 +1,26 @@
-package com.doddi.vulnerability_scanner_idea_plugin.service.scanners;
+package com.doddi.vulnerability_scanner_idea_plugin.service.scanners.nexusiq;
 
 import com.doddi.vulnerability_scanner_idea_plugin.config.UIConfiguration;
 import com.doddi.vulnerability_scanner_idea_plugin.service.ScannerService;
 
-public class OssIndexUIConfiguration
+public class NexusIQUIConfiguration
     implements UIConfiguration
 {
-  private static OssIndexUIConfiguration instance;
+  private static NexusIQUIConfiguration instance;
 
-  private OssIndexUIConfiguration() {
+  private NexusIQUIConfiguration() {
   }
 
-  public static OssIndexUIConfiguration getInstance() {
+  public static NexusIQUIConfiguration getInstance() {
     if (instance == null) {
-      instance = new OssIndexUIConfiguration();
+      instance = new NexusIQUIConfiguration();
     }
     return instance;
   }
 
   @Override
   public String name() {
-    return OssIndexScannerService.NAME;
+    return NexusIQScannerService.NAME;
   }
 
   @Override
@@ -30,16 +30,16 @@ public class OssIndexUIConfiguration
 
   @Override
   public boolean hasAppId() {
-    return false;
+    return true;
   }
 
   @Override
   public boolean hasRemoteUrl() {
-    return false;
+    return true;
   }
 
   @Override
   public Class<? extends ScannerService> scannerService() {
-    return OssIndexScannerService.class;
+    return NexusIQScannerService.class;
   }
 }
