+++ src/main/java/com/doddi/vulnerability_scanner_idea_plugin/service/scanners/ossindex/OssIndexScannerService.java
@@ -1,4 +1,4 @@
-package com.doddi.vulnerability_scanner_idea_plugin.service.scanners;
+package com.doddi.vulnerability_scanner_idea_plugin.service.scanners.ossindex;
 
 import java.util.Collections;
 import java.util.List;
