+++ src/main/java/com/doddi/vulnerability_scanner_idea_plugin/service/scanners/ossindex/OssIndexUIConfiguration.java
@@ -1,26 +1,26 @@
-package com.doddi.vulnerability_scanner_idea_plugin.service.scanners;
+package com.doddi.vulnerability_scanner_idea_plugin.service.scanners.ossindex;
 
 import com.doddi.vulnerability_scanner_idea_plugin.config.UIConfiguration;
 import com.doddi.vulnerability_scanner_idea_plugin.service.ScannerService;
 
-public class NexusIQUIConfiguration
+public class OssIndexUIConfiguration
     implements UIConfiguration
 {
-  private static NexusIQUIConfiguration instance;
+  private static OssIndexUIConfiguration instance;
 
-  private NexusIQUIConfiguration() {
+  private OssIndexUIConfiguration() {
   }
 
-  public static NexusIQUIConfiguration getInstance() {
+  public static OssIndexUIConfiguration getInstance() {
     if (instance == null) {
-      instance = new NexusIQUIConfiguration();
+      instance = new OssIndexUIConfiguration();
     }
     return instance;
   }
 
   @Override
   public String name() {
-    return NexusIQScannerService.NAME;
+    return OssIndexScannerService.NAME;
   }
 
   @Override
@@ -30,16 +30,16 @@ public class NexusIQUIConfiguration
 
   @Override
   public boolean hasAppId() {
-    return true;
+    return false;
   }
 
   @Override
   public boolean hasRemoteUrl() {
-    return true;
+    return false;
   }
 
   @Override
   public Class<? extends ScannerService> scannerService() {
-    return NexusIQScannerService.class;
+    return OssIndexScannerService.class;
   }
 }
