+++ src/main/resources/META-INF/plugin.xml
@@ -37,11 +37,11 @@
           serviceImplementation="com.doddi.vulnerability_scanner_idea_plugin.util.PluginInformationProvider"/>
 
       <applicationService
-          serviceImplementation="com.doddi.vulnerability_scanner_idea_plugin.service.scanners.OssIndexScannerService"/>
+          serviceImplementation="com.doddi.vulnerability_scanner_idea_plugin.service.scanners.ossindex.OssIndexScannerService"/>
       <applicationService
-          serviceImplementation="com.doddi.vulnerability_scanner_idea_plugin.service.scanners.NexusIQScannerService"/>
+          serviceImplementation="com.doddi.vulnerability_scanner_idea_plugin.service.scanners.nexusiq.NexusIQScannerService"/>
       <applicationService
-          serviceImplementation="com.doddi.vulnerability_scanner_idea_plugin.service.scanners.DummyScannerService"/>
+          serviceImplementation="com.doddi.vulnerability_scanner_idea_plugin.service.scanners.dummy.DummyScannerService"/>
     </extensions>
 
     <actions>
