+++ src/test/java/com/doddi/vulnerability_scanner_idea_plugin/system/MavenAnnotationTest.java
@@ -13,7 +13,7 @@ import com.doddi.vulnerability_scanner_idea_plugin.entity.LibraryToScan;
 import com.doddi.vulnerability_scanner_idea_plugin.entity.VulnerableComponent;
 import com.doddi.vulnerability_scanner_idea_plugin.service.ScannerManagerService;
 import com.doddi.vulnerability_scanner_idea_plugin.service.VulnerabilityService;
-import com.doddi.vulnerability_scanner_idea_plugin.service.scanners.DummyScannerService;
+import com.doddi.vulnerability_scanner_idea_plugin.service.scanners.dummy.DummyScannerService;
 import com.google.common.collect.ImmutableMap;
 import com.intellij.codeInsight.daemon.impl.HighlightInfo;
 import com.intellij.openapi.application.ApplicationManager;
